Fix flaky tests
Some tests were failing due to mismatch in Score in ZSetEntry.
Changed ZSetEntry Score to float64 type so that we can use
cmpopts.EquateApprox to allow for margin when comparing.

+++ client_test.go
@@ -58,7 +58,7 @@ func TestClient(t *testing.T) {
 						Retry:   defaultMaxRetry,
 						Queue:   "default",
 					},
-					Score: time.Now().Add(2 * time.Hour).Unix(),
+					Score: float64(time.Now().Add(2 * time.Hour).Unix()),
 				},
 			},
 		},
